+++ omero_figure/views.py
@@ -38,6 +38,7 @@ from omeroweb.webclient.decorators import login_required
 from omero_figure.scripts.omero.figure_scripts.Figure_To_Pdf import FigureExport, TiffExport
 
 from . import settings
+from . import utils
 
 try:
     from PIL import Image
@@ -112,8 +113,6 @@ def index(request, file_id=None, conn=None, **kwargs):
     and lay them out in canvas by dragging & resizing etc
     """
 
-    version = settings.OMERO_FIGURE_VERSION
-
     script_service = conn.getScriptService()
     sid = script_service.getScriptID(SCRIPT_PATH)
     script_missing = sid <= 0
@@ -121,7 +120,7 @@ def index(request, file_id=None, conn=None, **kwargs):
 
     context = {'scriptMissing': script_missing,
                'userFullName': user_full_name,
-               'version': version}
+               'version': utils.__version__}
     return render(request, "figure/index.html", context)