+++ src/js/views/figure_view.js
@@ -65,9 +65,12 @@
             this.listenTo(this.model, 'change:unsaved', this.renderSaveBtn);
             this.listenTo(this.model, 'change:figureName', this.renderFigureName);
 
+            // Full render if page_color changes (might need to update labels etc)
+            this.listenTo(this.model, 'change:page_color', this.render);
+            this.listenTo(this.model, 'change:page_color', this.renderPanels);
+
             // refresh current UI
             this.renderZoom();
-            // this.zoom_paper_to_fit();
 
             // 'Auto-render' on init.
             this.render();
@@ -167,7 +170,8 @@
             opts = {"PDF": "PDF",
                 "PDF & images": "PDF_IMAGES",
                 "TIFF": "TIFF",
-                "TIFF & images": "TIFF_IMAGES"};
+                "TIFF & images": "TIFF_IMAGES",
+                "to OMERO": "OMERO"};
             exportOption = opts[export_opt];
 
             // Get figure as json
@@ -611,7 +615,8 @@
 
         // Add a panel to the view
         addOne: function(panel) {
-            var view = new PanelView({model:panel});    // uiState:this.uiState
+            var page_color = this.model.get('page_color');
+            var view = new PanelView({model:panel, page_color:page_color});
             this.$figure.append(view.render().el);
         },
 
@@ -659,6 +664,15 @@
             }
         },
 
+        renderPanels: function() {
+            // Re-render all the panels...
+            // Remove and re-add
+            $('.imagePanel', this.$figure).remove();
+            this.model.panels.forEach(function(panel){
+                this.addOne(panel);
+            }.bind(this));
+        },
+
         // Render is called on init()
         // Update any changes to sizes of paper or canvas
         render: function() {
@@ -667,6 +681,7 @@
 
             var page_w = m.get('paper_width'),
                 page_h = m.get('paper_height'),
+                page_color = m.get('page_color'),
                 size = this.model.getFigureSize(),
                 canvas_w = Math.max(m.get('canvas_width'), size.w),
                 canvas_h = Math.max(m.get('canvas_height'), size.h),
@@ -691,7 +706,7 @@
                 $pages = $(".paper");
             }
 
-            $pages.css({'width': page_w, 'height': page_h});
+            $pages.css({'width': page_w, 'height': page_h, 'background-color': '#' + page_color});
 
             this.$figure.css({'width': size.w, 'height': size.h,
                     'left': figure_left, 'top': figure_top});