+++ src/js/views/lutpicker.js
@@ -50,6 +50,8 @@ var LutPickerView = Backbone.View.extend({
                 "orange_hot.lut",
                 "phase.lut",
                 "physics.lut",
+                "pup_br.lut",
+                "pup_nr.lut",
                 "rainbow_rgb.lut",
                 "red-green.lut",
                 "red_hot.lut",