+++ src/js/views/panel_view.js
@@ -28,6 +28,10 @@
             // During drag, model isn't updated, but we trigger 'drag'
             this.model.on('drag_resize', this.drag_resize, this);
 
+            // Used for rendering labels against page_color background
+            if (opts.page_color) {
+                this.page_color = opts.page_color;
+            }
             this.render();
         },
 
@@ -65,7 +69,7 @@
                         'height': h +'px'});
 
             // container needs to be square for rotation to vertical
-            $('.left_vlabels', this.$el).css('width', h + 'px');
+            $('.left_vlabels', this.$el).css('width', 3 * h + 'px');
 
             // update the img within the panel
             var zoom = this.model.get('zoom'),
@@ -143,11 +147,21 @@
             _.each(labels, function(l) {
                 // check if label is dynamic delta-T
                 var ljson = $.extend(true, {}, l);
+                // If label is same color as page (and is outside of panel)
+                if (ljson.color.toLowerCase() == self.page_color.toLowerCase() &&
+                        ["top", "bottom", "left", "right", "leftvert"].indexOf(l.position) > -1 ) {
+                    // If black -> white, otherwise -> black
+                    if (ljson.color === '000000') {
+                        ljson.color = 'ffffff';
+                    } else {
+                        ljson.color = '000000';
+                    }
+                }
                 if (typeof ljson.text == 'undefined' && ljson.time) {
                     ljson.text = self.model.get_time_label_text(ljson.time);
                 } else {
-                    // Escape all labels so they are safe
-                    ljson.text = _.escape(ljson.text);
+                    // Markdown also escapes all labels so they are safe
+                    ljson.text = markdown.toHTML(ljson.text);
                 }
                 positions[l.position].push(ljson);
             });
@@ -168,7 +182,7 @@
             self.$el.append(html);
 
             // need to force update of vertical labels layout
-            $('.left_vlabels', self.$el).css('width', self.$el.height() + 'px');
+            $('.left_vlabels', self.$el).css('width', 3 * self.$el.height() + 'px');
 
             return this;
         },