+++ src/js/views/right_panel_view.js
@@ -40,7 +40,7 @@
                 this.ctv.clear().remove();
             }
             if (selected.length > 0) {
-                this.ctv = new ChannelToggleView({models: selected});
+                this.ctv = new ImageDisplayOptionsView({models: selected});
                 $("#channelToggle").empty().append(this.ctv.render().el);
             }
             if (this.csv) {
@@ -249,6 +249,12 @@
         events: {
             "submit .new-label-form": "handle_new_label",
             "click .dropdown-menu a": "select_dropdown_option",
+            "click .markdown-info": "markdownInfo",
+        },
+
+        markdownInfo: function(event) {
+            event.preventDefault();
+            $("#markdownInfoModal").modal('show');
         },
 
         // Handles all the various drop-down menus in the 'New' AND 'Edit Label' forms
@@ -1492,10 +1498,10 @@
         }
     });
 
-    // Coloured Buttons to Toggle Channels on/off.
-    var ChannelToggleView = Backbone.View.extend({
+    // Options such as Rotation, Z-Projection etc.
+    var ImageDisplayOptionsView = Backbone.View.extend({
         tagName: "div",
-        template: JST["src/templates/channel_toggle_template.html"],
+        template: JST["src/templates/image_display_options_template.html"],
 
         initialize: function(opts) {
             // This View may apply to a single PanelModel or a list
@@ -1507,8 +1513,6 @@
         },
 
         events: {
-            "click .channel-btn": "toggle_channel",
-            "click .dropdown-menu a": "pick_color",
             "click .show-rotation": "show_rotation",
             "click .z-projection": "z_projection",
         },
@@ -1555,62 +1559,6 @@
             }
         },
 
-        pick_color: function(e) {
-            var color = e.currentTarget.getAttribute('data-color'),
-                $colorbtn = $(e.currentTarget).parent().parent(),
-                oldcolor = $(e.currentTarget).attr('data-oldcolor'),
-                idx = $colorbtn.attr('data-index'),
-                self = this;
-
-            if (color == 'colorpicker') {
-                FigureColorPicker.show({
-                    'color': oldcolor,
-                    'success': function(newColor){
-                        // remove # from E.g. #ff00ff
-                        newColor = newColor.replace("#", "");
-                        self.set_color(idx, newColor);
-                    }
-                });
-            } else if (color == 'lutpicker') {
-                FigureLutPicker.show({
-                    success: function(lutName){
-                        // LUT names are handled same as color strings
-                        self.set_color(idx, lutName);
-                    }
-                });
-            } else {
-                this.set_color(idx, color);
-            }
-            return false;
-        },
-
-        set_color: function(idx, color) {
-            if (this.models) {
-                this.models.forEach(function(m){
-                    m.save_channel(idx, 'color', color);
-                });
-            }
-        },
-
-        toggle_channel: function(e) {
-            var idx = e.currentTarget.getAttribute('data-index');
-
-            if (this.model) {
-                this.model.toggle_channel(idx);
-            } else if (this.models) {
-                // 'flat' means that some panels have this channel on, some off
-                var flat = $(e.currentTarget).hasClass('ch-btn-flat');
-                this.models.forEach(function(m){
-                    if(flat) {
-                        m.toggle_channel(idx, true);
-                    } else {
-                        m.toggle_channel(idx);
-                    }
-                });
-            }
-            return false;
-        },
-
         clear: function() {
             try {
                 this.$el.find('.rotation-slider').slider("destroy");
@@ -1619,7 +1567,7 @@
         },
 
         render: function() {
-            var json, html,
+            var html,
                 max_rotation = 0,
                 sum_rotation = 0,
                 sum_sizeZ = 0,
@@ -1628,60 +1576,18 @@
                 zp,
                 self = this;
             if (this.models) {
-
-                // Comare channels from each Panel Model to see if they are
-                // compatible, and compile a summary json.
-                json = [];
-                var compatible = true;
-
-                this.models.forEach(function(m){
-                    var chs = m.get('channels');
+                this.models.forEach(function(m, i){
                     rotation = m.get('rotation');
                     max_rotation = Math.max(max_rotation, rotation);
                     sum_rotation += rotation;
                     sum_sizeZ += m.get('sizeZ');
                     // start with a copy of the first image channels
-                    if (json.length === 0) {
-                        _.each(chs, function(c) {
-                            json.push($.extend(true, {}, c));
-                        });
+                    if (i === 0) {
                         z_projection = !!m.get('z_projection');
                     } else{
                         zp = !!m.get('z_projection');
                         if (zp !== z_projection) {
                             z_projection = undefined;
-                        }
-                        // compare json summary so far with this channels
-                        if (json.length != chs.length) {
-                            compatible = false;
-                        } else {
-                            // if attributes don't match - show 'null' state
-                            _.each(chs, function(c, cIndex) {
-                                if (json[cIndex].color != c.color) {
-                                    json[cIndex].color = 'ccc';
-                                }
-                                if (json[cIndex].active != c.active) {
-                                    json[cIndex].active = undefined;
-                                }
-                                // process the 'window' {min, max, start, end}
-                                var wdw = json[cIndex].window,    // the window we're updating
-                                    w = c.window;
-                                // if we haven't got a label yet, compare 'start' from 1st 2 panels
-                                if (typeof wdw.start_label === 'undefined') {
-                                    wdw.start_label = (w.start === wdw.start) ? parseInt(w.start,10) : '-';
-                                } else if (wdw.start_label != w.start) {
-                                    wdw.start_label = "-";      // otherwise revert to '-' unless all same
-                                }
-                                if (typeof wdw.end_label === 'undefined') {
-                                    wdw.end_label = (w.end === wdw.end) ? parseInt(w.end,10) : '-';
-                                } else if (wdw.end_label != w.end) {
-                                    wdw.end_label = "-";      // revert to '-' unless all same
-                                }
-                                wdw.min = Math.min(wdw.min, w.min);
-                                wdw.max = Math.max(wdw.max, w.max);
-                                wdw.start = wdw.start + w.start;    // average when done
-                                wdw.end = wdw.end + w.end;
-                            });
                         }
                     }
                 });
@@ -1697,15 +1603,7 @@
                 // if all panels have sizeZ == 1, don't allow z_projection
                 z_projection_disabled = (sum_sizeZ === this.models.length);
 
-                if (!compatible) {
-                    json = [];
-                }
-                // Add LUT offsets
-                json = json.map(function(ch){
-                    ch.lutBgPos = FigureLutPicker.getLutBackgroundPosition(ch.color);
-                    return ch;
-                });
-                html = this.template({'channels':json,
+                html = this.template({
                     'z_projection_disabled': z_projection_disabled,
                     'rotation': rotation,
                     'z_projection': z_projection});